Fix string comparison bug
The uniqued strings are not null-terminated. If a string's length is a
multiple of 8, the alignment will make it so that the string ends right
before the length of the next string in the binary dump, so it will be
terminated by the zero bytes in that length field, leading to bad
comparisons.